# Enable parallel execution for `make_digits`