<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰
v    Before hitting that submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Description

On MacBook Air, [4 performance cores, 4 efficiency cores](https://developer.apple.com/news/?id=vk3m204o)
**Before:**
```
MSM for Bls12_381::G1
                        time:   [273.98 ms 290.22 ms 312.05 ms]
```

**After:**
```
MSM for Bls12_381::G1
                        time:   [253.65 ms 255.12 ms 256.76 ms]
                        change: [-18.328% -12.097% -6.8345%] (p = 0.00 < 0.05)
                        Performance has improved.
```

---

On 16 cores / 32 threads @ 4.2 GHz
**Before:**
```
MSM for Bls12_381::G1
                        time:   [126.18 ms 126.47 ms 126.87 ms]
```
**After:**
```
MSM for Bls12_381::G1
                        time:   [114.80 ms 115.20 ms 115.68 ms]
                        change: [-9.3244% -8.9102% -8.4379%] (p = 0.00 < 0.05)
                        Performance has improved.
```

---

Before we can merge this PR, please make sure that all the following items have been
checked off. If any of the checklist items are not applicable, please leave them but
write a little note why.

- [ ] Targeted PR against correct branch (master)
- [ ] Linked to GitHub issue with discussion and accepted design OR have an explanation in the PR that describes this work.
- [ ] Wrote unit tests
- [ ] Updated relevant documentation in the code
- [ ] Added a relevant changelog entry to the `Pending` section in `CHANGELOG.md`
- [ ] Re-reviewed `Files changed` in the GitHub PR explorer